split up and simplified assets middleware
