+++ node/lib/assets.js
@@ -1,50 +1,17 @@
 const logger = require('@financial-times/n-logger').default;
-const denodeify = require('denodeify');
 const path = require('path');
-const fs = require('fs');
-const readFile = denodeify(fs.readFile);
 const nPolyfillIo = require('@financial-times/n-polyfill-io');
-const chokidar = require('chokidar');
 const nUiManager = require('./n-ui-manager');
-
-function constructLinkHeader (hashedAssets) {
-	return function (file, meta, opts) {
-		meta = meta || {};
-		opts = opts || {};
-		const header = [];
-		header.push(`<${opts.hashed ? hashedAssets.get(file) : file }>`);
-		Object.keys(meta).forEach(key => {
-			header.push(`${key}="${meta[key]}"`)
-		});
-
-		if (!meta.rel) {
-			header.push('rel="preload"')
-		}
-
-		header.push('nopush');
-		this.append('Link', header.join('; '))
-	}
-}
+const linkHeaderFactory = require('./link-header');
+const stylesheetManager = require('./stylesheet-manager');
+const messages = require('./messages');
 
 module.exports = function (options, directory, hashedAssets) {
 
-	const localAppShell = process.env.NEXT_APP_SHELL === 'local';
-	if (localAppShell) {
-		logger.warn(`
-/*********** n-express warning ************/
+	const useLocalAppShell = process.env.NEXT_APP_SHELL === 'local';
 
-You have set the environment variable NEXT_APP_SHELL=local
-This should only be used if you are actively developing
-n-ui/n-html-app within the context of an app (by bower linking
-or similar). It will slow down your build A LOT and be a slightly
-less accurate approximation of the production app!!!!
-
-If you do not need this behaviour run
-
-			unset NEXT_APP_SHELL
-
-/*********** n-express warning ************/
-`);
+	if (useLocalAppShell) {
+		logger.warn(messages.APP_SHELL_WARNING);
 	}
 	// Attempt to retrieve the json file used to configure n-ui
 	let nUiConfig;
@@ -53,104 +20,57 @@ If you do not need this behaviour run
 	} catch (e) {}
 
 
-	const headCsses = options.withHeadCss ? fs.readdirSync(`${directory}/public`)
-		.filter(name => /^head[\-a-z]*\.css$/.test(name))
-		.map(name => [name, fs.readFileSync(`${directory}/public/${name}`, 'utf-8')])
-		.reduce((currentHeadCsses, currentHeadCss) => {
-			currentHeadCsses[currentHeadCss[0].replace('.css', '')] = currentHeadCss[1];
-			return currentHeadCsses;
-		}, {}) : {};
-
-	if (process.NODE_ENV !== 'production') {
-		const paths = Object.keys(headCsses).map(css => `${directory}/public/${css}.css`);
-		chokidar.watch(paths)
-			.on('change', (path) => {
-				readFile(path, 'utf-8').then((content) => {
-					const name = path.match(/\/(head.*).css$/)[1];
-					headCsses[name] = content;
-					logger.info(`Reloaded head CSS: ${name}`);
-				});
-			})
-			.on('unlink', (path) => {
-				const name = path.match(/\/(head.*).css$/)[1];
-				delete headCsses[name];
-				logger.info(`Deleted head CSS: ${name}`);
-				logger.warn('Please note you will need to restart app if you add new head CSS files');
-			});
-	}
-
-	const linkHeader = constructLinkHeader(hashedAssets);
+	const stylesheets = stylesheetManager.getStylesheets(options, directory);
+	const linkHeader = linkHeaderFactory(hashedAssets);
+	const nUiUrlRoot = nUiManager.getUrlRoot(hashedAssets);
 
 	return (req, res, next) => {
 
-		const nUiUrlRoot = nUiManager.getUrlRoot(hashedAssets);
 		// This middleware relies on the presence of res.locals.flags.
 		// In some scenarios (e.g. using handlebars but not flags) this
 		// won't be present
-		res.locals.flags = res.locals.flags || {};
+		const flags = res.locals.flags || {};
 
-		const swCriticalCss = req.get('ft-next-sw') && res.locals.flags.swCriticalCss;
 		// define a helper for adding a link header
 		res.linkResource = linkHeader;
 
 		// backwards compatible - can remove once n-ui templates updated everywhere
-		res.locals.headCsses = headCsses;
+		res.locals.stylesheets = stylesheets;
 		if (req.accepts('text/html')) {
 			res.locals.javascriptBundles = [];
 			res.locals.cssBundles = [];
 			res.locals.criticalCss = [];
+			res.locals.nUiConfig = nUiConfig;
 
 			// work out which assets will be required by the page
-			if (options.withNUiJsBundle) {
-				res.locals.nUiConfig = nUiConfig;
-				res.locals.javascriptBundles.push(
-					`${nUiUrlRoot}es5${(res.locals.flags.nUiBundleUnminified || localAppShell ) ? '' : '.min'}.js`
-				);
-				res.locals.javascriptBundles.push(hashedAssets.get('main-without-n-ui.js'));
-			}
-			else {
-				res.locals.javascriptBundles.push(hashedAssets.get('main.js'));
-			}
-
-			// output the default link headers just before rendering
-			const originalRender = res.render;
-			const polyfillRoot = `//${res.locals.flags.polyfillQA ? 'qa.polyfill.io' : 'next-geebee.ft.com/polyfill'}/v2/polyfill.min.js`;
+			const polyfillRoot = `//${flags.polyfillQA ? 'qa.polyfill.io' : 'next-geebee.ft.com/polyfill'}/v2/polyfill.min.js`;
 
 			res.locals.polyfillCallbackName = nPolyfillIo.callbackName;
 			res.locals.polyfillUrls = {
-				enhanced: polyfillRoot + nPolyfillIo.getQueryString({
-					enhanced: true,
-					withRum: res.locals.flags.polyfillsRUM ? 1 : 0
-				}),
-				core: polyfillRoot + nPolyfillIo.getQueryString({
-					enhanced: false
-				})
+				enhanced: polyfillRoot + nPolyfillIo.getQueryString({enhanced: true}),
+				core: polyfillRoot + nPolyfillIo.getQueryString({enhanced: false})
 			}
 
-			res.locals.javascriptBundles.push(res.locals.polyfillUrls.enhanced);
+			res.locals.javascriptBundles.push(
+				`${nUiUrlRoot}es5${(flags.nUiBundleUnminified || useLocalAppShell ) ? '' : '.min'}.js`,
+				hashedAssets.get('main-without-n-ui.js'),
+				res.locals.polyfillUrls.enhanced
+			);
+
+			// output the default link headers just before rendering
+			const originalRender = res.render;
 
 			res.render = function (template, templateData) {
 
 				let cssVariant = templateData.cssVariant || res.locals.cssVariant;
-				cssVariant = cssVariant ? '-' + cssVariant : '';
+				cssVariant = cssVariant ? `-${cssVariant}` : '';
 
 				// define which css to output in the critical path
 				if (options.withHeadCss) {
-					if (`head-n-ui-core` in headCsses) {
-						if (swCriticalCss) {
-							res.locals.cssBundles.push({
-								path: `${nUiUrlRoot}main.css`
-							});
-						} else {
-							// even if the page hasn't been loaded via sw, we still want to encourage the browser to preload
-							// this shared critical css file for the next visit. Support for prefetch is wider than for preload too
-							if (res.locals.flags.swCriticalCss) {
-								res.linkResource(`${nUiUrlRoot}main.css`, {as: 'style', rel: 'prefetch'});
-							}
-							res.locals.criticalCss.push(headCsses[`head-n-ui-core`])
-						}
+					if (`head-n-ui-core` in stylesheets) {
+						res.locals.criticalCss.push(stylesheets[`head-n-ui-core`])
 					}
-					res.locals.criticalCss.push(headCsses[`head${cssVariant}`]);
+					res.locals.criticalCss.push(stylesheets[`head${cssVariant}`]);
 				}
 
 				res.locals.cssBundles.push({
